pyproject.toml
[project]
name = "text-insight-clustering-service"
version = "0.1.0"
description = "Text insight clustering service for take-home task"
requires-python = ">=3.11"

dependencies = [
  # Core
  "pydantic>=2,<3",
  "numpy>=2.3.0",

  # Semantic embedding & clustering
  "scikit-learn>=1.8",

  # Sentiment
  "vaderSentiment>=3.3",

  # LLM access (OpenAI-compatible)
  "requests>=2.32",

  # Testing
  "pytest>=9.0",

  # ENV
  "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
  "isort>=7.0.0",
  "pytest>=9.0",
  "ruff>=0.14.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.isort]
profile = "black"
src_paths = ["src", "tests", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Ignore line length (handled by formatter if used)

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]  # Allow imports after sys.path modification in scripts
